Added a .gitignore file with IntelliJ, Angular, and Python support.
For anyone who doesn't know, the .gitignore file (as the file name hints) ignores all the files/folders listed in the file when pushing files to the git. This is helpful, as many files are either created on build or are specific to someone's machine. Feel free to add more to the .gitignore if you need. There are really simple tutorials/generators online if needed.