Added a navigation bar to the top of the home page.
I tried for hours to add the navigation bar and settled on using ng-bootstrap, which is Bootstrap 4 (current alpha version) written in Angular.
Other ways that I tried that did not work for one reason or another:
* Bootstrap 3.6.6 ([routerLink] was not understood)
* Angular Material (Couldn't figure out how to get the components to work properly)

This commit still has Angular Material leftovers in it, but any calls to it should be commented out.

TODO: Completely remove Angular Material remnents.